requirements.txt
python-telegram-bot==21.10
yt-dlp==2024.12.23
requests==2.32.3
PyYAML==6.0.2
mutagen==1.47.0